% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_NetVAR_plot.R
\name{plot.NetVAR}
\alias{plot.NetVAR}
\title{Plotting function for objects of class NetVAR}
\usage{
\method{plot}{NetVAR}(x, join_graphs = TRUE, scale_range = TRUE, legend = FALSE, ...)
}
\arguments{
\item{x}{An object of \code{class} NetVAR}

\item{join_graphs}{Logical. If TRUE, then heatmaps for each lagged 
parameter matrix are returned in one plot. If FALSE, a single figure per 
heatmap is generated.}

\item{scale_range}{Logical. If TRUE, parameters larger 1 or smaller -1 are 
truncated to 1 and -1 purely for visualisation purposes. If FALSE, then the 
actual values are plotted in the heatmaps.}

\item{legend}{Add a legend to plots. Default FALSE}

\item{...}{Further arguments, currently none.}
}
\value{
None
}
\description{
Plots an object of class NetVAR
}
