% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_NetVAR_summary.R
\name{summary.NetVAR}
\alias{summary.NetVAR}
\title{Printing function for objects of class NetVAR}
\usage{
\method{summary}{NetVAR}(object, ...)
}
\arguments{
\item{object}{An object of \code{class} NetVAR}

\item{...}{Further arguments to be passed over. Currently none.}
}
\value{
None
}
\description{
\code{\link{summary.NetVAR}} prints the values of an object of class NetVAR.
}
