% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TradingData}
\alias{TradingData}
\title{Trading data for 11 US stocks from the finance sector}
\format{
A matrix with 5 minute trading data for 11 stocks on 22.02.2022.
}
\source{
Yahoo-Finance.
}
\description{
A dataset containing the 5 minute log returns of 11 US stocks for the 
22.02.2022.
}
